+++ aviasales-common/src/main/java/ru/aviasales/common/dto/request/SalesCategoryReq.java
@@ -24,6 +24,6 @@ public class SalesCategoryReq {
 
     @NotNull(message = "Процент комиссии по умолчанию не может быть пустым")
     @Schema(description = "Процент комиссии по умолчанию")
-    private Double defaultCommissionPercent;
+    private Long defaultCommissionPercent;
 
 }
